Refactor Package Cache Test case
