+++ language-server/modules/langserver-compiler/src/test/java/org/ballerinalang/langserver/compiler/LSPackageCacheTest.java
@@ -7,41 +7,26 @@ import org.ballerinalang.langserver.compiler.workspace.repository.WorkspacePacka
 import org.ballerinalang.model.elements.PackageID;
 import org.ballerinalang.repository.PackageRepository;
 import org.testng.Assert;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import org.wso2.ballerinalang.compiler.util.CompilerContext;
 import org.wso2.ballerinalang.compiler.util.Name;
 import org.wso2.ballerinalang.compiler.util.Names;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Set;
 
 public class LSPackageCacheTest {
 
-    private static final Path ROOT_DIR = com.google.common.io.Files.createTempDir().toPath();
-
-    private static final String TESTS_SAMPLES = "src" + File.separator + "test" + File.separator + "resources"
-            + File.separator + "source";
-
-    protected static final Path SAMPLES_COPY_DIR = ROOT_DIR.resolve("source");
-
-    @BeforeMethod
-    public void loadTestCases() throws IOException {
-        File source = new File(TESTS_SAMPLES);
-        File destination = SAMPLES_COPY_DIR.toFile();
-        destination.mkdirs();
-        org.apache.commons.io.FileUtils.copyDirectory(source, destination);
-    }
+    private static final Path RES_DIR = Paths.get("src/test/resources/").toAbsolutePath();
 
     @Test
     public void testInvalidate() throws Exception {
-        Path filePath = SAMPLES_COPY_DIR.resolve("singlepackage").resolve("io-sample.bal");
+        Path filePath = RES_DIR.resolve("source").resolve("singlepackage").resolve("io-sample.bal");
         compileFileAndCheckCache(filePath);
-        filePath = SAMPLES_COPY_DIR.resolve("multipackages").resolve("sample").resolve("main.bal");
+        filePath = RES_DIR.resolve("source").resolve("multipackages").resolve("sample").resolve("main.bal");
         compileFileAndCheckCache(filePath);
     }
 
@@ -69,10 +54,5 @@ public class LSPackageCacheTest {
         PackageID nPackageIDAnon = new PackageID(Names.ANON_ORG, new Name(pkgName), Names.DEFAULT_VERSION);
         Assert.assertFalse(packageMayKeySet.contains(nPackageIDDemo.bvmAlias()));
         Assert.assertFalse(packageMayKeySet.contains(nPackageIDAnon.bvmAlias()));
-    }
-
-    @AfterClass
-    public void cleanSamplesCopy() throws IOException {
-        org.apache.commons.io.FileUtils.deleteDirectory(SAMPLES_COPY_DIR.toFile());
     }
 }
