+++ language-server/modules/langserver-core/src/test/resources/definition/.ballerina/readme.txt
@@ -1,0 +1,1 @@
+resides within the .ballerina file
